Set up authentication

+++ routes/bands.js
@@ -1,6 +1,7 @@
 const path = require("path")
 const router = require('express').Router();
 const db = require('../models');
+const isAuthenticated = require("../config/middleware/isAuthenticated");
 
 router.route('/bands').get((req, res) => {
     db.Band.findAll({})
@@ -10,26 +11,30 @@ router.route('/bands').get((req, res) => {
     .catch(err => res.json(err));
 });
 
-router.delete("/bands:id", (req, res) => {
-    db.Band.destroy({
-        where: {
-            id: req.params.id
-        }
-    }).then(res.send(200))
-    .catch(err => res.json(err))
+router.delete("/bands:id", isAuthenticated, (req, res) => {
+    if(req.user.name === 'rootroot') {
+        db.Band.destroy({
+            where: {
+                id: req.params.id
+            }
+        }).then(res.send(200))
+        .catch(err => res.json(err))
+    }
 });
 
-router.post('/bands', (req, res) => {
+router.post('/bands', isAuthenticated, (req, res) => {
     const band = req.body;
-    db.Band.create({
-        band: band.band,
-        year: band.year,
-        day: band.day,
-        month: band.month,
-        start: band.start,
-        end: band.end
-    }).then(response => res.json(response))
-    .catch(err => res.json(err))
+    if(req.user.name === 'rootroot') {
+        db.Band.create({
+            band: band.band,
+            year: band.year,
+            day: band.day,
+            month: band.month,
+            start: band.start,
+            end: band.end
+        }).then(response => res.json(response))
+        .catch(err => res.json(err))
+    }
 })
 
 module.exports = router;
